fix(test): add sript for running the tests in CI
Given that we use Cypress to run our tests, the fact that it needs the
webserver to be running, and it's considered bad practice to just start
it from within the tests, a separate script to run from within the CI
environment is needed.

The `start-server-and-test` package handles starting the server and
waits for it to respond before executing the test. You should really
look at its great docs, but the `ci` script uses one of it aliases to
start the server and then run the `test` script, and since
`create-react-app` comes with `webpack-dev-server` and it doesn't
respond to a HEAD request, we need to send explicitly a GET request.