[SKIP CI] fix(README): use new all-contributors badge
According to [1], the pattern used by the all-contributors bot to find
and update the badge makes impossible inlining the badges due to the way
comments are rendered in markdown. In response to that, in [2] a new(er)
badge that doesn't need to be manually updated by the bot was made.
Still, from the status of [3] it seems the cli (which apparently powers the
API used by the bot) has yet to be updated to use it.

Fixes #41

[1] all-contributors/all-contributors#361
[2] all-contributors/all-contributors#406
[3] all-contributors/all-contributors-cli#266